Enh #7555: yii\log\Logger refactoring
